move config/file and llm to envs
