+++ src/server/globalConfig/genServerAiProviderConfig.ts
@@ -4,7 +4,7 @@ import { extractEnabledModels, transformToAiModelList } from '@lobechat/utils';
 import * as AiModels from 'model-bank';
 import { AiFullModelCard } from 'model-bank';
 
-import { getLLMConfig } from '@/config/llm';
+import { getLLMConfig } from '@/envs/llm';
 
 interface ProviderSpecificConfig {
   enabled?: boolean;
