+++ src/server/globalConfig/index.ts
@@ -1,8 +1,8 @@
 import { authEnv } from '@/config/auth';
-import { fileEnv } from '@/config/file';
 import { enableNextAuth } from '@/const/auth';
 import { isDesktop } from '@/const/version';
 import { appEnv, getAppConfig } from '@/envs/app';
+import { fileEnv } from '@/envs/file';
 import { knowledgeEnv } from '@/envs/knowledge';
 import { langfuseEnv } from '@/envs/langfuse';
 import { parseSystemAgent } from '@/server/globalConfig/parseSystemAgent';
