+++ src/server/modules/ModelRuntime/index.ts
@@ -1,7 +1,7 @@
 import { ModelProvider, ModelRuntime } from '@lobechat/model-runtime';
 import { ClientSecretPayload } from '@lobechat/types';
 
-import { getLLMConfig } from '@/config/llm';
+import { getLLMConfig } from '@/envs/llm';
 
 import apiKeyManager from './apiKeyManager';
