+++ src/services/file/_deprecated.test.ts
@@ -1,8 +1,8 @@
 import { Mock, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { fileEnv } from '@/config/file';
 import { FileModel } from '@/database/_deprecated/models/file';
 import { DB_File } from '@/database/_deprecated/schemas/files';
+import { fileEnv } from '@/envs/file';
 import { clientS3Storage } from '@/services/file/ClientS3';
 import { serverConfigSelectors } from '@/store/serverConfig/selectors';
 import { createServerConfigStore } from '@/store/serverConfig/store';
