+++ src/services/upload.ts
@@ -4,7 +4,7 @@ import { uuid } from '@lobechat/utils';
 import dayjs from 'dayjs';
 import { sha256 } from 'js-sha256';
 
-import { fileEnv } from '@/config/file';
+import { fileEnv } from '@/envs/file';
 import { edgeClient } from '@/libs/trpc/client';
 import { API_ENDPOINTS } from '@/services/_url';
 import { clientS3Storage } from '@/services/file/ClientS3';
